// src/app/models/columnSetting.ts
export class columnSetting {
	rnId: boolean;
	firstName: boolean; 
	lastName: boolean;
	email: boolean;
	cohortDate: boolean;
	employmentStartDate: boolean;
	typeOfUnit: boolean;
	age: boolean;
	gender: boolean;
	ethnicity: boolean;
	education: boolean;
	degreeReceived: boolean;
	previousNursingDegree: boolean;
	overallGPA: boolean;
	meetingPlaceAccess: boolean;
	leave: boolean;
	termination: boolean;
	terminationUpdatedBy: boolean;
	typeofLeave: boolean;
	dateTerminated: boolean;
	leaveFromDate: boolean;
	terminationReason: boolean;
	leaveToDate: boolean;
	terminationLastUpdatedOn: boolean;
}